refactor(forms): consolidate types and update payment slider config
- Move form types from forms.type.ts to types.ts
- Add PaymentSliderConfig constants to config
- Update PaymentForm to use new config values
- Remove unused DatePicker import from PaymentForm
